Fix sequence items visibility with strictforward case and non-problems
